major renaming operation!!!
renamed file and corresponding links, commenting and so on and so
forth. layouts have a "-view" & "-edit" mode. This so I can more easily
develop how the audience sees it and how the owner sees it. Then for
practical purposes the "edit-" layouts have a temporary container
colour through class="edit"