alert-usab and alert-bug classes added
The purpose of this is to quickly show in the UI where things need to
be done.
alert-usab is about links that do not yet work or go to the wrong
places or stuff that is simply unclear.
alert-bug is about something that technically doesnt work or works in
an unexpected manner